refactor WhatToExpect Component to folder
